Files re-structure
Prev: appointment-list component was combining calendar view and appointment view(list-view) which does not make sense.
Now: appointment component combines calendar view and appointment-list view.
I was working on wrong file name.
