+++ angular-app/src/app/pages/appointments/appointment-list/appointment-list.component.css
@@ -1,22 +1,20 @@
 /*
-    Created on : Apr 27, 2022, 12:26:03 PM
+    Created on : Apr 28, 2022, 12:10:25 PM
     Author     : Matt Crow <mattcrow19@gmail.com>
-
-    updated by: Rojan Maharjan 10/06/2022
-
 */
 
-.appt-list {
+.appointment {
     display: flex;
-    flex-direction: column;
     align-items: center;
-    gap: 10px;
+    justify-content: center;
+    flex-direction: column;
+    width: 100%;
 }
 
-.table-view{
-  width: 800px;
+.appointment table{
+width: 100%;
 }
 
-.calender-view{
-  width:800px;
+mat-paginator{
+  width: inherit;
 }
